refresh site data in background when cache expiration is hit
Previously, when cached site data expired, it would block the current request (and all other requests) on refreshing the site data (eg by downloading a new repository archive zip). Now it does that in the background. If the refresh fails, it clears the cache so that the problem is made user-visible on the next request (so that monitoring tools will detect the problem and the site won't continue silently serving stale data).

This will significantly improve page load times because users won't have to wait for the site data to be fetched each 5 min.