disable serving other versions of content
It is rarely used to visit https://docs.sourcegraph.com/@someotherrevision/somepath. But this feature adds a lot of complexity and has caused memory leaks and other problems.

This commit just disables the functionality. Removal will come later. I wanted to disable it for now because it's fast to do so, and so then we can remove it later knowing that it definitely wasn't still in use.