Check versioned requests that 404 for possible redirects
At present, versioned requests do not get parsed for redirects, which
causes issues when we have /help links in Sourcegraph that depend on a
redirect being in place. As redirects are defined on the asset VFS, we
can't trivially check the redirects for that specific version, but as a
fallback, we can see if there are any current redirects that might help.

This addresses (but doesn't actually fix) sourcegraph/sourcegraph#17931.