Add support for more metadata fields (#75)
Add more fields to metadata struct, enabling use front matter to set open graph tags for SEO. It partly addresses #65.

Design-wise, all new fields are hardcoded in the metadata struct. No error will be raised if they aren't present. But any key not referenced by the Metadata struct will raise an error.

A plain old meta field in the front matter could have been a good choice, but my primary concern with that is that have no way to ensure that those fields are properly populated, which would result in believing that the pages are properly annotated while they are not, which is difficult to spot. There are ways to fix this, but this requires additional work.