Make rel links checker more precise.
The predicate used to detect a link pointing to "foo" where the actual
page is at "foo/index.md" was too narrow and prevented to link to other
filetypes such as .csv.

It nows check if the linked file has an extension rather than this
extension just being ".md", which enables to link to CSV files for
example.